fixup! fixup! move ElasticsearchTaskHandler to the provider package
